Fix featured blog post title wrapping issues
The featured blog post title wouldn't wrap if it was too long, and
instead would be cut off. Since there's no easy way to restrict the
length of the title in WordPress, this isn't acceptable, as it too often
results in a broken blog front page.

Instead, anchor the h2 and paragraph excerpt tags to the bottom of the
featured image and allow them to flow up.